+++ packages/components/src/molecules/LearnPage/LearnPage.stories.tsx
@@ -1,0 +1,13 @@
+import {ComponentMeta, ComponentStory} from '@storybook/react';
+import LearnPage from './LearnPage';
+
+export default {
+  title: 'Molecules/LearnPage',
+  component: LearnPage,
+} as ComponentMeta<typeof LearnPage>;
+
+const Template: ComponentStory<typeof LearnPage> = args => (
+  <LearnPage {...args} />
+);
+
+export const MainLearnPage = Template.bind({});